+++ custom_components/chargeamps/light.py
@@ -5,7 +5,7 @@ import logging
 from homeassistant.components.light import SUPPORT_BRIGHTNESS, LightEntity
 
 from . import ChargeampsEntity
-from .const import DOMAIN, DOMAIN_DATA
+from .const import DOMAIN, DOMAIN_DATA, SCAN_INTERVAL
 
 _LOGGER = logging.getLogger(__name__)